# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger UI",
        "description": "L5 Swagger API description",
        "contact": {
            "name": "Dinh Huan",
            "email": "dinhhuan1357@gmail.com"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/combo/create": {
            "post": {
                "tags": [
                    "Combo"
                ],
                "summary": "Tạo combo",
                "description": "Tạo combo mới với ảnh",
                "operationId": "ccbb32065afaa93805583dd933a770a9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "ComboName",
                                    "Description",
                                    "Price",
                                    "Cb_Sale",
                                    "Image",
                                    "CategoryID"
                                ],
                                "properties": {
                                    "ComboName": {
                                        "type": "string",
                                        "example": "Combo Dark Test + Vitamin D3 220 viên - Combo Tăng Test, Tăng Đề Kháng"
                                    },
                                    "Description": {
                                        "type": "string",
                                        "example": "ombo bao gồm :Znutrition Dark Test - Tăng Testosterone mạnh mẽ Nature Made Vitamin D3 2000iu - 220 Viên"
                                    },
                                    "Price": {
                                        "type": "integer",
                                        "example": 14000000
                                    },
                                    "Cb_Sale": {
                                        "type": "integer",
                                        "example": 15
                                    },
                                    "CategoryID": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "Image": {
                                        "description": "Ảnh combo (1 ảnh)",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tạo combo thành công"
                    },
                    "500": {
                        "description": "Tạo combo thất bại"
                    }
                }
            }
        },
        "/api/combo/add": {
            "post": {
                "tags": [
                    "Combo"
                ],
                "summary": "Thêm sản phẩm",
                "description": "Thêm sản phẩm vào combo",
                "operationId": "b21da60f9a913299aa7e7489eb81b347",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "ComboID",
                                    "ProductID",
                                    "VariantID",
                                    "Quantity"
                                ],
                                "properties": {
                                    "ComboID": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "ProductID": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "VariantID": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "Quantity": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Thêm sản phẩm thành công"
                    },
                    "500": {
                        "description": "Thêm sản phẩm thất bại"
                    }
                }
            }
        },
        "/api/combo/{id}": {
            "get": {
                "tags": [
                    "Combo"
                ],
                "summary": "Tìm thông tin combo",
                "description": "Tìm thông tin chi tiết combo có {id}",
                "operationId": "aa2616f4261d4e80148c0b84d71f51eb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id của combo",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tìm thấy combo"
                    },
                    "500": {
                        "description": "Lỗi dịch vụ"
                    }
                }
            }
        },
        "/api/combo/{id}/products": {
            "get": {
                "tags": [
                    "Combo"
                ],
                "summary": "thông tin sản phẩm trong combo",
                "description": "Tìm thông tin sản phẩm thuộc combo",
                "operationId": "525cb9c18641abf897ebad6bee1047d1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id của combo cần tìm sản phẩm",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tìm sản phẩm thành công"
                    },
                    "400": {
                        "description": "Tìm sản phẩm thất bại"
                    },
                    "500": {
                        "description": "Lỗi dịch vụ"
                    }
                }
            }
        },
        "/api/combo/{combo_id}": {
            "delete": {
                "tags": [
                    "Combo"
                ],
                "summary": "Xóa combo có {id}",
                "description": "xóa combo",
                "operationId": "15b466f8c6f14da6e500364c107ec666",
                "parameters": [
                    {
                        "name": "combo_id",
                        "in": "path",
                        "description": "combo_id  cần xóa",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Xóa thành công"
                    },
                    "400": {
                        "description": "Xóa thất bại"
                    },
                    "500": {
                        "description": "Lỗi dịch vụ"
                    }
                }
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "sanctum": []
        }
    ]
}